requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
djangorestframework==3.13.1
djangorestframework-gis @ https://github.com/openwisp/django-rest-framework-gis/tarball/master
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.2
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
